# backend/cmd/server/docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "API for managing financial data, chat interactions, and scenario planning",
        "title": "Financial Chat System API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "email": "support@example.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
...
        "/v2/cashflow/incomes": {
            "get": {
                "security": [
                    {
                        "SessionID": []
                    },
                    {
                        "AuthToken": []
                    }
                ],
                "description": "Returns paginated incomes for the authenticated user",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Incomes V2"
                ],
                "summary": "List incomes (v2)",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Max items to return (-1 for all)",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Number of items to skip",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "SessionID": []
                    },
                    {
                        "AuthToken": []
                    }
                ],
                "description": "Creates a new income for the authenticated user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Incomes V2"
                ],
                "summary": "Create an income (v2)",
                "parameters": [
                    {
                        "description": "Income data",
                        "name": "income",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.incomeV2CreateInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/financial-chat-system_backend_internal_financial_v2_repository.Income"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "SessionID": []
                    },
                    {
                        "AuthToken": []
                    }
                ],
                "description": "Bulk deletes all incomes for the authenticated user in a single query.",
                "tags": [
                    "Bulk Delete V2"
                ],
                "summary": "Delete all incomes (v2)",
                "operationId": "deleteBulkIncomes",
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/v2/cashflow/incomes/{id}": {
            "put": {
                "security": [
                    {
                        "SessionID": []
                    },
                    {
                        "AuthToken": []
                    }
                ],
                "description": "Updates an income with versioning support",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Incomes V2"
                ],
                "summary": "Update an income (v2)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Income ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Income data",
                        "name": "income",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.incomeV2Input"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/financial-chat-system_backend_internal_financial_v2_repository.Income"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "SessionID": []
                    },
                    {
                        "AuthToken": []
                    }
                ],
                "description": "Deletes an income and all descendant versions",
                "tags": [
                    "Incomes V2"
                ],
                "summary": "Delete an income (v2)",
                "operationId": "deleteIncome",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Income ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/v2/cashflow/incomes/{id}/stop": {
            "post": {
                "security": [
                    {
                        "SessionID": []
                    },
                    {
                        "AuthToken": []
                    }
                ],
                "description": "Sets the endDate on an income (soft delete)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Incomes V2"
                ],
                "summary": "Stop an income (v2)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Income ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Stop input with endDate",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.stopInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/financial-chat-system_backend_internal_financial_v2_repository.Income"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "financial-chat-system_backend_internal_financial_v2_repository.Income": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "number"
                },
                "category": {
                    "type": "string"
                },
                "cpfWageType": {
                    "description": "'ow' (Ordinary Wages) or 'aw' (Additional Wages)",
                    "type": "string"
                },
                "endDate": {
                    "description": "NULL means ongoing",
                    "type": "string"
                },
                "frequency": {
                    "type": "string"
                },
                "growthRate": {
                    "type": "number"
                },
                "growthStrategy": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "impactFrequency": {
                    "description": "How often delta adds (NULL for base/override)",
                    "type": "string"
                },
                "impactKind": {
                    "description": "NULL = base item, 'delta' = additive, 'override' = replaces",
                    "type": "string"
                },
                "incomeType": {
                    "description": "CPF-related fields",
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "notes": {
                    "type": "string"
                },
                "parentId": {
                    "type": "string"
                },
                "personId": {
                    "description": "FK to persons table (required)",
                    "type": "string"
                },
                "personName": {
                    "description": "Display name from persons table (read-only, populated via JOIN)",
                    "type": "string"
                },
                "scenarioEventId": {
                    "description": "Scenario impact fields",
                    "type": "string"
                },
                "startDate": {
                    "description": "Precise start date (day-level) - now required",
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
...
        "handlers.incomeV2CreateInput": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "string"
                },
                "category": {
                    "type": "string"
                },
                "cpfWageType": {
                    "type": "string"
                },
                "endDate": {
                    "type": "string"
                },
                "frequency": {
                    "type": "string"
                },
                "growthRate": {
                    "type": "string"
                },
                "growthStrategy": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "notes": {
                    "type": "string"
                },
                "personId": {
                    "description": "Required FK to persons table",
                    "type": "string"
                },
                "startDate": {
                    "type": "string"
                }
            }
        },
        "handlers.incomeV2Input": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "string"
                },
                "category": {
                    "type": "string"
                },
                "frequency": {
                    "type": "string"
                },
                "growthRate": {
                    "type": "string"
                },
                "growthStrategy": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "notes": {
                    "type": "string"
                },
                "parentId": {
                    "type": "string"
                },
                "personId": {
                    "description": "Required FK to persons table",
                    "type": "string"
                },
                "startDate": {
                    "type": "string"
                },
                "updateMode": {
                    "type": "string"
                }
            }
        },
...
        "handlers.stopInput": {
            "type": "object",
            "properties": {
                "endDate": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "AuthToken": {
            "description": "JWT token for authentication (HMAC-signed with BACKEND_SHARED_SECRET). In dev mode, use GET /api/dev/token?user_id=\u003cuuid\u003e to generate one.",
            "type": "apiKey",
            "name": "X-Auth-Token",
            "in": "header"
        }
    }
}